fix solr deploy

+++ config/rubber/deploy-solr.rb
@@ -29,13 +29,15 @@ namespace :rubber do
             update-alternatives --install "/usr/bin/java" "java" "/usr/lib/jvm/jdk1.7/bin/java" 1
             update-alternatives --install "/usr/bin/javac" "javac" "/usr/lib/jvm/jdk1.7/bin/javac" 1
             update-alternatives --install "/usr/bin/javaws" "javaws" "/usr/lib/jvm/jdk1.7/bin/javaws" 1
-
+          fi
+          if [ ! -d "#{rubber_env.tomcat_dest_folder}" ]; then
             echo 'installing tomcat'
             curl -o /tmp/apache-tomcat-7.0.34.tar.gz http://ftp.heanet.ie/mirrors/www.apache.org/dist/tomcat/tomcat-7/v7.0.39/bin/apache-tomcat-7.0.39.tar.gz
             sudo mkdir -p #{rubber_env.tomcat_dest_folder}
             tar -zxf /tmp/apache-tomcat-7.0.34.tar.gz -C #{rubber_env.tomcat_dest_folder}
             rm /tmp/apache-tomcat-7.0.34.tar.gz
-
+          fi
+          if [ ! -d "#{rubber_env.solr_home_dest_foler}" ]; then
             echo 'installing solr'
             curl -o /tmp/apache-solr-4.0.0.tgz http://ftp.heanet.ie/mirrors/www.apache.org/dist/lucene/solr/3.6.2/apache-solr-3.6.2.tgz
             tar -zxf /tmp/apache-solr-4.0.0.tgz -C /tmp
